tests/integration/requirements-test.txt
coverage[toml]==7.2.5
pytest==7.3.1
PyYAML==6.0.1
json
tenacity==8.2.3